mua funcionando, error con la parada en los semaforos

+++ src/movil/tiposVehiculos/Bus.scala
@@ -1,7 +1,7 @@
 package movil.tiposVehiculos
 import movil._
 import puntos._
-class Bus(placa:String)(posicion:Punto,velocidad:Velocidad,aceleracion:Int) extends Vehiculo(placa)(posicion,velocidad,aceleracion){
+class Bus(placa:String)(posicion:Punto,velocidad:Velocidad,aceleracion:Double) extends Vehiculo(placa)(posicion,velocidad,aceleracion){
 }
 
 object Bus {
