+++ src/movil/tiposVehiculos/Carro.scala
@@ -2,7 +2,7 @@ package movil.tiposVehiculos
 import movil._
 import puntos._
 
-class Carro(placa:String)(posicion:Punto,velocidad:Velocidad,aceleracion:Int) extends Vehiculo(placa)(posicion,velocidad,aceleracion){
+class Carro(placa:String)(posicion:Punto,velocidad:Velocidad,aceleracion:Double) extends Vehiculo(placa)(posicion,velocidad,aceleracion){
 }
 object Carro {
   def generarPlaca : String = {
